refactor(network): abstract LISH want/search serving into own module
